input: Add new hook DeviceSendEventsProc for XTest
For Xwayland, we need to be able to send the events that would normally
be processed by the normal Xserver event processing to be forwarded to
the Wayland compositor (somehow).

Add a new hook “DeviceSendEventsProc” attached to the device so that
Xwayland can implement its own routine instead of the “normal” XTest
implementation which generate and process X input events.

Signed-off-by: Olivier Fourdan <ofourdan@redhat.com>